move voicelist command function
